requirements.txt
flask~=3.0.2
python-dotenv~=1.0.1
requests~=2.31.0
grpcio~=1.62.1
grpcio-tools~=1.62.1
flask-cors~=4.0.0